Successfully scraped some URLs from vox.com/news!
My scrapy settings are super confused right now. I should look into setting this
up as a scrapy project. I should probably spend more time with the scrapy
tutorial too.

The Coloradoan's html is much harder to parse, and my requests seem to be
redirected to a https://accountsolution.gcion.com/ site, which I think is the
Coloradoan trying to get me to subscribe. The last spider I ran just looped on
these redirects, which would seem to indicate that my free reads ran out. Bottom
line: the Coloradoan is not a good website for learning and testing.